Implement recursive deletion inside kio_file. Brings down the 13-seconds testcase to 2 seconds, although it kills progress information completely for the case of "deleting one directory with stuff in it". BUG: 174144
svn path=/trunk/KDE/kdelibs/; revision=927756

+++ kioslave/file/file_unix.cpp
@@ -529,7 +529,8 @@ void FileProtocol::del(const KUrl& url, bool isfile)
        *****/
 
       kDebug( 7101 ) << "Deleting directory " << url.url();
-
+      if (!deleteRecursive(path))
+          return;
       if ( ::rmdir( _path.data() ) == -1 ) {
 	if ((errno == EACCES) || (errno == EPERM))
 	  error(KIO::ERR_ACCESS_DENIED, path);
